I'm in the process of minimizing references to PopNode, so that there will be less work to do when I replace PopNode with Segment for the non-simulation version of legofit and legosim. There are no longer PopNode references in sampndx.c. I've changed popnodetab.{c,h} into strptrmap.{c,h}, which are also free of PopNode references.
I also fixed some run-time errors involving strhash, in which I was truncating
long unsigned values to unsigned.

+++ src/misc.c
@@ -436,7 +436,7 @@ char       *strlowercase(char *s) {
 }
 
 /// Hash a character string
-unsigned strhash(const char *ss) {
+unsigned long strhash(const char *ss) {
     unsigned long hashval;
     int c;
     const unsigned char *s = (const unsigned char *) ss;
